map class name refactor
