+++ client/src/map_canvas.rs
